discord_bot/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
selenium>=4.0.0
beautifulsoup4>=4.9.3
webdriver-manager>=3.8.0
pandas>=1.3.0  # For the calendar functionality 